Reform to include employer payroll taxes in taxable income (#5635)
* Created new variables for employer contributions; finished new formula for irs_gross_income

* Finished formula for not taxing social security; added boilerplate for reforms

* Update tax_payroll_not_benefits_reform.py

* Update changelog_entry.yaml

* fixed tax_employer_payroll.py

fixes are:
1. renaming yaml containing boolean to "in_effect"
2. adding a five-year-forward check

also sorry that i renamed the directory. i didn't realize that it breaks all github tracked changes.

i'll add the unit tests in a subsequent commit.

* fixed line lengths

* more formatting

i had the black library make the reformatting changes.

* fixed issue i think

this is a path and not a custom object, so it shouldn't have a parenthesis

* more line wrapping

* more line wrapping

* Commit before switching to update parameter and adding unit tests

* mid-commit

* switched to parametric reform

adding unit test next

* unit tests added

(the one for the reform is incomplete because it will be deleted)

* removed reform unit test

no unit test for purely parametric reforms

* formatting

* formatting again

* retrieve income sources list before adding new sources

(1) The function that feeds into modify_parameters now appends to the existing income sources list, rather than replaces it. (2) Added a unit test for the reform as well. (3) Directories and files are renamed. (4) Added a missing return statement to the function that feeds into modify_parameters.

* Update policyengine_us/tests/policy/contrib/crfb/tax_employer_payroll_tax.yaml

* Update policyengine_us/tests/policy/contrib/crfb/tax_employer_payroll_tax.yaml

* Update policyengine_us/tests/policy/contrib/crfb/tax_employer_payroll_tax.yaml

---------

Co-authored-by: Max Ghenis <mghenis@gmail.com>